* worked on templates + introduced widget classes and views to reduce clutter
