+++ public/modules/autoload.php
@@ -11,10 +11,12 @@ require_once('activity/Activity.php'); // (*)
 require_once('activity/Comment.php'); // (*)
 require_once('activity/Activity_Shout.php'); // (*)
 require_once('activity/Activity_Event.php'); // (*)
+require_once('activity/Activity_Event_Widgets.php'); // (*)
 require_once('activity/Activity_Poll.php'); // (*)
 require_once('activity/Activity_ActionMsg.php'); // (*)
 require_once('pages/Pages.php'); // (*)
 require_once('gw2api/gw2api.php'); // (*)
+require_once('gw2api/gw2api_Widgets.php'); // (*)
 /*
  * for development purposes
  */
