+++ public/themes/boilerplate/css/main.css
@@ -110,6 +110,7 @@ body {
 .footer-container {
     background: #EDA73E;
 }
+
 aside {
     background: #ffffff;
     color: #EA6C12;
@@ -122,10 +123,6 @@ aside ul {
 
 aside ul li {
     clear: both;
-}
-
-aside time {
-    float: right;
 }
 
 .title {
@@ -506,25 +503,54 @@ nav.activities li:hover div.action:hover {
 
 /* End of "Micro clearfix" */
 
+aside a {
+    text-decoration: none;
+}
 
+aside a:hover {
+    text-decoration: underline;
+}
+
+.widget table,
 .module table {
     width: 100%;
 }
 
+.widget table tr th,
 .module table tr th {
     color: #ffffff;
     background-color: #EDA73E;
 }
 
+.widget table tr:nth-child(2n),
 .module table tr:nth-child(2n) {
     color: #EA6C12;
     background-color: #ffffff;
 }
 
+.widget table td a,
+.module table td a {
+    color: black;
+    text-decoration: none;
+}
+
+.widget table td a:hover,
+.module table td a:hover {
+    text-decoration: underline;
+}
+
+.widget table tr:nth-child(2n) td a,
+.module table tr:nth-child(2n) td a {
+    color: #EA6C12;
+    background-color: white;
+}
+
+.widget table td.center,
 .module table td.center {
     text-align:center;
 }
 
+.widget table td.right,
 .module table td.right {
     text-align: right;
 }
@@ -543,6 +569,27 @@ span[type="validation"] {
     clear: both;
     background-color: #ffcccc;
 }
+
+section.widget ul {
+    margin: 0; padding: 0;
+    list-style-type: none; 
+}
+
+section.widget ul li {
+    margin: 0; padding: 0;
+}
+
+aside section.widget h3 {
+    display: none;
+    margin: 0; padding: 0;
+    font-size: 16px;
+    color: #eda73e;
+}
+
+aside section.widget time {
+    float: right;
+}
+
 
 /* ==========================================================================
    Media Queries
