--- public/themes/boilerplate/views/activity/event/activity_signup_schedule_view.php
@@ -1,8 +1,0 @@
-{##signups_loop##}
-<article class="activity">
-    <section class="content">
-        {##activity_title##} @ {##activity_event_date##} {##activity_event_time##}
-        <a href="{##activity_details_link##}">{##activity_details_link_text##}</a>
-    </section>
-</article>
-{/##signups_loop##}
