--- public/themes/boilerplate/views/activity/event/scheduled_activities.php
@@ -1,10 +1,0 @@
-{##activity_message##}
-<ul>
-{##activity_loop##}
-<li class="event">
-    <a href="{##activity_details_link##}">
-        <span>{##activity_title##}</span> : <time datetime="{##activity_event_date##} {##activity_event_time##}">{##activity_event_date##} @ {##activity_event_time##}</time>
-    </a>
-</li>
-{/##activity_loop##}
-</ul>
